Move create_fake_incident() to better place
